Update job handlers to support manual accounts
