+++ client/src/App.tsx
@@ -4,6 +4,7 @@ import { AccountCard } from './components/AccountCard'
 import { BalanceCard } from './components/BalanceCard'
 import { ChainCard } from './components/ChainCard'
 import { Layout } from './components/Layout'
+import { PortfolioCard } from './components/PortfolioCard'
 import { TokenCard } from './components/TokenCard'
 import { Home } from './screens/Home'
 
@@ -12,10 +13,11 @@ function App() {
     <Layout>
       <Routes>
         <Route path="/" element={<Home />} />
-        <Route path="/portfolio" element={<BalanceCard />} />
+        <Route path="/portfolio" element={<PortfolioCard />} />
         <Route path="/chains" element={<ChainCard />} />
         <Route path="/accounts" element={<AccountCard />} />
         <Route path="/tokens" element={<TokenCard />} />
+        <Route path="/balances" element={<BalanceCard />} />
       </Routes>
     </Layout>
   )
